+++ lib/modules/home/widgets/property_grid_view.dart
@@ -163,7 +163,7 @@ class _PropertyGridViewState extends ConsumerState<PropertyGridView> {
     });
   }
 
-  formatBudget(double budget) {
+  formatBudget(int budget) {
     if (budget < 10000000) {
       return "${(budget / 100000).toStringAsFixed(2)} L";
     } else {
